server/requirements.txt
alembic==1.15.2
aniso8601==10.0.1
backports-datetime-fromisoformat==2.0.3
bcrypt==4.3.0
bidict==0.23.1
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Faker==37.4.0
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
gunicorn==21.2.0
h11==0.16.0
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
pytz==2024.1
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.1
Werkzeug==3.1.3
wsproto==1.2.0
